MidiTrack.Merge(), MidiTrack.TrackName fix, MidiEvent.Owner, MidiEvent.AbsoluteTime & OnNoteVoiceMidiEvent.Duration
MidiTrack.Merge() will take the input track and place all of its events into the current track, except for the track end event.
MidiTrack.TrackName property will now return the correct name if the track contains more than one name event.

MidiEvent.Owner property is the MidiTrack that the event belongs to.
MidiEvent.AbsoluteTime property is the amount of ticks it takes to reach this event.

OnNoteVoiceMidiEvent.Duration property is the amount of ticks this note stays on.